Alta de tareas de mantenimiento - Modificaciones para tener tarea por km y por tiempo en las entidades.

+++ 2016AD_G04_Repositorio/src/main/java/com/ADG04/bean/Vehiculo/DTO_TareasPorTiempo.java
@@ -2,17 +2,17 @@ package com.ADG04.bean.Vehiculo;
 
 public class DTO_TareasPorTiempo extends DTO_TareaMantenimiento{
 	private static final long serialVersionUID = 1L;
-	private float cantidadDias;
+	private int cantidadDias;
 
 	public DTO_TareasPorTiempo(){
 		
 	}
 
-	public float getCantidadDias() {
+	public int getCantidadDias() {
 		return cantidadDias;
 	}
 
-	public void setCantidadDias(float cantidadDias) {
+	public void setCantidadDias(int cantidadDias) {
 		this.cantidadDias = cantidadDias;
 	}
